requirements.txt
python-telegram-bot==13.15
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Pillow==10.1.0
openpyxl==3.1.2
requests==2.32.3